Fix link if file does not exist

+++ link.py
@@ -7,5 +7,6 @@ dotfiles = manifest.read().splitlines()
 for dotfile in dotfiles:
     source_file = '{}/{}'.format(os.getcwd(), dotfile)
     target_file = '{}/{}'.format(os.path.expanduser('~'), dotfile)
-    os.remove(target_file)
+    if os.path.isfile(target_file):
+        os.remove(target_file)
     os.symlink(source_file, target_file)
